backend/requirements.txt
# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Encryption & Security
pycryptodome==3.19.0
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy>=2.0.35
alembic>=1.13.0
# psycopg2-binary==2.9.9  # Commented out for local dev (using SQLite)

# QR Code & File Handling
qrcode[pil]==7.4.2
python-multipart==0.0.6

# CORS & Middleware
# python-cors==1.0.0  # Not needed - using FastAPI's built-in CORS

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
